BREAKING: Refactor functionality for v2.0.0 release
- Refactored core components in Solteq Tand for improved maintainability and performance
- Updated pyproject.toml to version 2.0.0 reflecting the breaking changes
