+++ mbu_dev_shared_components/solteqtand/__init__.py
@@ -1,0 +1,32 @@
+"""
+This module provides the main entry point for the Solteq Tand application.
+"""
+from .application import (
+    app_handler,
+    appointment,
+    base_ui,
+    clinic,
+    document,
+    edi_portal,
+    event,
+    patient,
+    exceptions,
+    helper_functions,
+    SolteqTandApp
+)
+from .database import SolteqTandDatabase
+
+__all__ = [
+    "app_handler",
+    "appointment",
+    "base_ui",
+    "clinic",
+    "document",
+    "edi_portal",
+    "event",
+    "patient",
+    "exceptions",
+    "helper_functions",
+    "SolteqTandApp",
+    "SolteqTandDatabase"
+]
