+++ mbu_dev_shared_components/solteqtand/application/__init__.py
@@ -1,0 +1,1 @@
+from .app_handler import SolteqTandApp
