+++ mbu_dev_shared_components/solteqtand/application/helper_functions.py
