+++ mbu_dev_shared_components/solteqtand/database/__init__.py
@@ -1,0 +1,1 @@
+from .db_handler import SolteqTandDatabase
